Images to be used
Well download you have the normal which will be used when multithreading
becomes available and then you have the inverted which is when you go to
download.
You have the notHere which is the 404, you also have the two setting for
when it is selected or not. Finally you have the update button. all of
the buttons are 50px by 50px